requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
tortoise-orm==0.20.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
python-multipart==0.0.6
aiosqlite
black==23.12.1
boto3==1.34.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2